Deletion of Zaeri Light theme
I feel like Light themes are just a joke and just here to hurt your eyes on purpose. However, I maybe will do one to try in the future. But for the moment, I'll stick to dark theme and maybe try High Contrast one even if I don't really feel confortable with it.